pyproject.toml
[tool.poetry]
name = "tria-bot"
version = "0.1.0"
description = ""
authors = ["Lucas Lucyk <llucyk@grupospec.com>"]
readme = "README.md"
packages = [{include = "tria_bot"}]


[tool.poetry.dependencies]
python = "^3.10"
anyio = "^4.0.0"
aiormq = "^6.7.7"
orjson = "^3.9.10"
redis-om = "^0.2.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
